:worker: change calibration matrix of av1

+++ pose_graph/src/parameters.h
@@ -35,7 +35,7 @@
     extern std::string POSE_GRAPH_SAVE_NAME;
     extern int ROW;
     extern int COL;
-    extern double ransac_error;
+    extern float ransac_error;
     extern int gps_init;
     extern std::string RESULT_PATH;
     extern int DEBUG_IMAGE;
